api/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
psycopg2-binary 
pandas
questionnaire
pytest
